arm_pack_manager - Adds an dump parts command
